% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptBin_rox.R
\name{acceptinterval}
\alias{acceptinterval}
\title{acceptinterval}
\usage{
acceptinterval(x, n, level = 0.95, tolerance = 1e-04)
}
\arguments{
\item{x}{}

\item{n}{}

\item{level}{}

\item{tolerance}{}
}
\description{
acceptinterval
}
\examples{
acceptinterval (x =  , n =  , level = 0.95, tolerance = 1e-04)
}
