% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobFun.R
\name{blob}
\alias{blob}
\title{blob}
\usage{
blob(tf, est = "est", upper = upper, lower = lower)
}
\arguments{
\item{tf}{Dataframe with parameter estimate, CI upper and lower limits.}
}
\description{
blob uses the upper bound when above median, or lower bound when below
}
