% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPlotR-package.R
\docType{package}
\name{checkPlotR-package}
\alias{checkPlotR}
\alias{checkPlotR-package}
\title{checkPlotR: Tools for visually assessing nominal CI and p-values}
\description{
Statistically valid p-values should have uniform density\; frequentist confidence intervals (CI) should achieve (at least) nominal coverage, and be unbiased, i.e. probability that the target value falls below the CI is equal to the probablility that the target value falls above the CI. These properties are guaranteed with pivotal statistics for continuous distributions, but not for many statistical applications of p-values and CI. In such cases p-values and CI may be visually tested with checkplots (which assess the uniformity of p-values) and with rangeplots, a family of plots to vizualize the distribution of CI under Monte-Carlo sampling.
}
\author{
\strong{Maintainer}: Michael Roswell \email{mike.roswell@gmail.com} (\href{https://orcid.org/0000-0002-8479-9184}{ORCID})

Authors:
\itemize{
  \item Jonathan Dushoff \email{jdushoff@gmail.com} (\href{https://orcid.org/0000-0003-0506-4794}{ORCID})
  \item Michael Li \email{lim88@mcmaster.ca}
}

}
\keyword{internal}
