% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milliFun.R
\name{milli}
\alias{milli}
\title{Order estimates for a millipede plot}
\usage{
milli(tf)
}
\arguments{
\item{tf}{Dataframe with parameter estimate, CI upper and lower limits.}
}
\description{
The Millipede plot orders CI by the [mean] parameter estimate
}
