% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multBinom.R
\name{multBinom}
\alias{multBinom}
\title{compute p-values, CI for binomial proportions}
\usage{
multBinom(dat, prob0, n, testv = c("binom.test", "accept", "chisq", "wald"))
}
\arguments{
\item{dat}{integer vector, observed number of successes}

\item{prob0}{scalar; true probability of success}

\item{n}{scalar, number of trials}

\item{testv}{character, which binomial test to use for estimating CI and p}
}
\value{
dataframe same length as \code{dat} with p-value estimates, upper and lower
confidence limits, and the name of the test used
}
\description{
A wrapper for several frequentist tests for binomial proportions, it provides
consistent outputs regardless of test method.
}
\examples{

#set parameters
p_true <- 0.7
n_trials <- 10
n_samples <- 10

#generate data
mydat <- rbinom(n_samples, n_trials, p_true)

run tests
multBinom(mydat, ptrue, n_trials, testv = "binom.test")
multBinom(mydat, ptrue, n_trials, testv = "wald")

}
