% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFuns.R
\name{rangePlot}
\alias{rangePlot}
\title{Create Rangeplot Based On Order Function}
\usage{
rangePlot(
  tf,
  target = mean(tf$est),
  orderFun = slug,
  conf = 0.95,
  opacity = 0.2,
  fatten = 0.1,
  title = "Range plot",
  targ_num = 1000
)
}
\description{
Confidence intervals (CI) consisting of an estimate, upper, and lower limit
from Monte-Carlo sampling (given that the statistical hypothesis is true) are
arranged according to \code{orderFun} and compared with the hypothesized
(true) value of the statistic to examine statistical coverage and bias of CI
}
\details{
@template tf_template
@param target scalar, value of statistic under (true) hypothesis
@param orderFun one of \code{c(slug, milli, blob)}, determining the way CI
are arranged in the plot
@param conf scalar in \link{0,1}, nominal confidence level of simulated CI
@param opacity scalar value for \code{alpha}
@param fatten scalar passed to \code{geom_pointrange}
@param title character vector for ggtitle
@param targ_num desired number of lines on graph, for dynamic data thinning

@return a ggplot object with simulated CI arranged about their target value

@export
}
