% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugFun.R
\name{slug}
\alias{slug}
\title{slug}
\usage{
slug(tf)
}
\arguments{
\item{tf}{Dataframe with parameter estimate, CI upper and lower limits.}
}
\description{
slug orders by the bound you care about, dynamically
}
