% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangePlot.R
\name{rangePlot}
\alias{rangePlot}
\title{Create Rangeplot Based On Order Function}
\usage{
rangePlot(
  tf,
  target = "sampleMean",
  orderFun = slug,
  conf = 0.95,
  opacity = 0.2,
  fatten = 0.1,
  title = "Range plot",
  targNum = 1000
)
}
\arguments{
\item{tf}{Dataframe with parameter estimate, CI upper and lower limits.}

\item{target}{scalar, value of statistic under (true) hypothesis, or
"sampleMean", default, which calculates the sample mean.}

\item{orderFun}{one of \code{c(slug, milli, blob)}, determining the way CI
are arranged in the plot}

\item{conf}{scalar in [0,1], nominal confidence level of simulated CI}

\item{opacity}{scalar value for \code{alpha}}

\item{fatten}{scalar passed to \code{geom_pointrange}}

\item{title}{character vector for ggtitle}

\item{targNum}{desired number of lines on graph, for dynamic data thinning}
}
\value{
a ggplot object with simulated CI arranged about their target value
}
\description{
Confidence intervals (CI) consisting of an estimate, upper, and lower limit
from Monte-Carlo sampling (given that the statistical hypothesis is true) are
arranged according to \code{orderFun} and compared with the hypothesized
(true) value of the statistic to examine statistical coverage and bias of CI
}
